Add files via upload
Car price prediction is a machine learning project that involves predicting the selling price of a car based on various features like make, model, year, mileage, fuel type, and other specifications. This project is useful for both buyers and sellers in the automotive market to estimate the fair market value of a car. It involves data collection, preprocessing, feature engineering, model training, evaluation, and deployment.